Expand example to show differences with raw pointers
extract<const T&> requires an exact match with T so that the
existing installed object can be referenced. In the case of
extract<const shared_ptr<A>&>, where the Python
object holds a shared_ptr<B>, extract would need to create a
new shared_ptr<A> copy from the original that would not be
a reference to existing held instance so it fails.
Extracting by value allows a copy of the shared_ptr to be created,
using the alias constructor, and hence extaction suceeds.